+++ .build.ps1
@@ -42,8 +42,8 @@ task GitStatus -If (Test-Path .git) {
 	}
 }
 
-# Copy Invoke-Build.ps1 and help from working location to the project.
-# Fail (assert) if the project file is newer, it is not supposed to be.
+# Copy scripts and help from working location to the project.
+# Fail if the project files are newer, to be resolved manually.
 task UpdateScript {
 	$target = Get-Item Invoke-Build.ps1 -ErrorAction 0
 	$source = Get-Item (Get-Command Invoke-Build.ps1).Definition
@@ -58,8 +58,7 @@ task UpdateScript {
 
 # Build the PowerShell help file.
 task Help {
-	$script = (Get-Command Invoke-Build.ps1).Definition
-	$dir = Split-Path $script
+	$dir = Split-Path (Get-Command Invoke-Build.ps1).Definition
 	. Helps.ps1
 	Convert-Helps Invoke-Build.ps1-Help.ps1 $dir\Invoke-Build.ps1-Help.xml
 }
@@ -71,22 +70,20 @@ task Package ConvertMarkdown, Help, UpdateScript, GitStatus, {
 	$null = mkdir z\tools
 
 	# copy project files
-	Copy-Item -Destination z\tools -Recurse @(
-		'Demo'
-		'Build.ps1'
-		'Invoke-Build.ps1'
-		'LICENSE.txt'
-	)
+	Copy-Item -Destination z\tools -Recurse `
+	Demo,
+	Build.ps1,
+	Invoke-Build.ps1,
+	LICENSE.txt
 
 	# move generated files
-	Move-Item -Destination z\tools @(
-		'Invoke-Build.ps1-Help.xml'
-		'README.htm'
-		'Release-Notes.htm'
-	)
+	Move-Item -Destination z\tools `
+	Invoke-Build.ps1-Help.xml,
+	README.htm,
+	Release-Notes.htm
 }
 
-# Make the package, test in it with Demo renamed to [ ] and clean.
+# Make the package, test in it with Demo renamed to [ ], clean.
 #! It is not supposed to be called with other tasks.
 task PackageTest Package, {
 	# make it much harder with square brackets
@@ -111,11 +108,11 @@ task Zip Package, {
 # Make the NuGet package.
 task NuGet Package, {
 	$text = @'
-Invoke-Build.ps1 is a build automation tool implemented as a standalone
-PowerShell script. It invokes tasks defined in build scripts written in
-PowerShell with a few domain-specific language features. Build flow and
-concepts are similar to MSBuild. Scripts are similar to psake but look more
-like usual due to standard PowerShell parameters and script scope variables.
+Invoke-Build.ps1 is a build and test automation tool implemented as a single
+PowerShell script. It invokes tasks defined in scripts written in PowerShell
+with domain-specific language features. Build flow and concepts are similar to
+MSBuild. Scripts are similar to psake but look more like usual due to standard
+script parameters and script scope variables.
 '@
 	# nuspec
 	Set-Content z\Package.nuspec @"
@@ -131,15 +128,15 @@ like usual due to standard PowerShell parameters and script scope variables.
 		<requireLicenseAcceptance>false</requireLicenseAcceptance>
 		<summary>$text</summary>
 		<description>$text</description>
-		<tags>powershell build automation</tags>
+		<tags>powershell build automation testing</tags>
 	</metadata>
 </package>
 "@
 	# pack
-	exec { NuGet pack z\Package.nuspec -NoDefaultExcludes }
+	exec { NuGet pack z\Package.nuspec -NoDefaultExcludes -NoPackageAnalysis }
 }
 
-# Make both zip and NuGet packages
+# Make zip and NuGet packages.
 task Pack Zip, NuGet
 
 # Calls tests infinitely to be sure it works and nothing leaks.
@@ -155,12 +152,12 @@ task Loop {
 # Invoke-Build-Test.log in %TEMP%. Then compare it with a new output file in
 # this directory. If they are the same then remove the new file. Otherwise
 # start %MERGE%.
-# * After the test call UpdateScript.
+# * After tests call UpdateScript.
 task Test {
-	# invoke tests, get the output and result
+	# invoke tests, get output and result
 	$output = Invoke-Build . Demo\.build.ps1 -Result result | Out-String -Width:9999
 
-	assert ($result.AllTasks.Count -eq 143) $result.AllTasks.Count
+	assert ($result.AllTasks.Count -eq 141) $result.AllTasks.Count
 	assert ($result.Tasks.Count -eq 29) $result.Tasks.Count
 
 	assert ($result.AllErrorCount -eq 28) $result.AllErrorCount