+++ Demo/ProtectedTasks.build.ps1
@@ -5,10 +5,6 @@
 
 .Example
 	Invoke-Build . ProtectedTasks.build.ps1
-
-.Link
-	Invoke-Build
-	.build.ps1
 #>
 
 # Import tasks Error1 and Error2 (dot-sourced because imported with data).